Refactor robot detection into a class
